requirements.txt
Flask
openai
pymongo
elasticsearch
python-dotenv
